decomp UpdateSound and label more music stuff
