+++ ld_script.txt
@@ -72,8 +72,6 @@ SECTIONS {
         src/file_system.o(.text);
         src/main.o(.text);
         asm/code_800B5F0.o(.text);
-        src/music_pre.o(.text);
-        asm/music.o(.text);
         src/music.o(.text);
         src/code_800C9CC.o(.text);
         src/cpu.o(.text);
